Added PWM Function For Motor Speed
Wrote the first version of a function to control the motor speed. Using Timer0 (8 bit) set in CTC mode so that the output pins will toggle on the provided time period. Ability to select left or right for finer control of motors during tuning. *Note: Prescaler for timer needs to be set after some math to determine speed and frequency of motors.